Remove unecesary methods to open an URL & File
